# Configurable averageBits