@fd0 please review this PR addressing the #15. I haven't noticed any significant performance changes:

```
➜  chunker git:(UCS-2912) ✗ go version
go version go1.9 linux/amd64

➜  chunker git:(UCS-2912) ✗ go test -bench=BenchmarkChunker -benchtime=10s ./...
goos: linux
goarch: amd64
pkg: github.com/vitalyisaev2/chunker
BenchmarkChunkerWithSHA256-4   	     100	 169943162 ns/op	 197.45 MB/s
--- BENCH: BenchmarkChunkerWithSHA256-4
BenchmarkChunker-4             	     200	  78779123 ns/op	 425.93 MB/s
--- BENCH: BenchmarkChunker-4
PASS
ok  	github.com/vitalyisaev2/chunker	42.818s

➜  chunker git:(master) go test -bench=BenchmarkChunker -benchtime=10s ./...
goos: linux
goarch: amd64
pkg: github.com/restic/chunker
BenchmarkChunkerWithSHA256-4   	     100	 170337016 ns/op	 196.99 MB/s
--- BENCH: BenchmarkChunkerWithSHA256-4
BenchmarkChunker-4             	     200	  78613392 ns/op	 426.83 MB/s
--- BENCH: BenchmarkChunker-4
PASS
ok  	github.com/restic/chunker	42.485s
```