@vitalyisaev2 2017-10-19
I have also added `ResetWithBoundaries` method, because one could need to reset min / max chunksize settings along with read buffer and polynomial.