@fd0 2017-10-19
Nice, the performance is indeed the same:
```
$ benchstat before after
name                 old time/op   new time/op   delta
ChunkerWithSHA256-4    190ms ± 3%    188ms ± 0%   ~     (p=0.690 n=5+5)
Chunker-4             88.2ms ± 0%   87.8ms ± 1%   ~     (p=0.095 n=5+5)
name                 old speed     new speed     delta
ChunkerWithSHA256-4  177MB/s ± 3%  178MB/s ± 0%   ~     (p=0.690 n=5+5)
Chunker-4            380MB/s ± 0%  382MB/s ± 1%   ~     (p=0.095 n=5+5)
```